tests: move init tests and change generated directory
