+++ .gitignore
@@ -54,4 +54,3 @@ packages/**/*.map
 
 # temporary test files
 test-assets/
-
